Pass 3A

+++ src/kvraft/common.go
@@ -18,7 +18,7 @@ type PutAppendArgs struct {
 	Key   string
 	Value string
 	Op    string // "Put" or "Append"
-	Idx   int
+	Idx   uint64
 	// You'll have to add definitions here.
 	// Field names must start with capital letters,
 	// otherwise RPC will break.
